# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DesafioBRQ, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Candidato": {
      "post": {
        "tags": [
          "Candidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Candidato"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Candidato/id/{id}": {
      "get": {
        "tags": [
          "Candidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Candidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Candidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Candidato/cpf/{cpf}": {
      "get": {
        "tags": [
          "Candidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cpf",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Certificacao": {
      "post": {
        "tags": [
          "Certificacao"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCertificacaoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Certificacao"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Certificacao/id/{id}": {
      "get": {
        "tags": [
          "Certificacao"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Certificacao"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCertificacaoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Certificacao"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CertificacaoCandidato": {
      "post": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCertificacaoCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CertificacaoCandidato/id/{id}": {
      "get": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCertificacaoCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CertificacaoCandidato/candidato/{idCandidato}": {
      "get": {
        "tags": [
          "CertificacaoCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idCandidato",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Skill": {
      "post": {
        "tags": [
          "Skill"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateSkillDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Skill"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Skill/id/{id}": {
      "get": {
        "tags": [
          "Skill"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Skill"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateSkillDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Skill"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/SkillCandidato": {
      "post": {
        "tags": [
          "SkillCandidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateSkillCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "SkillCandidato"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/SkillCandidato/id/{id}": {
      "get": {
        "tags": [
          "SkillCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "SkillCandidato"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateSkillCandidatoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "SkillCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/SkillCandidato/candidato/{idCandidato}": {
      "get": {
        "tags": [
          "SkillCandidato"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idCandidato",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateCandidatoDTO": {
      "required": [
        "cpf",
        "email",
        "genero",
        "nascimento",
        "nome",
        "telefone"
      ],
      "type": "object",
      "properties": {
        "cpf": {
          "format": "int64",
          "maximum": 99999999999,
          "minimum": 11111111111,
          "type": "integer"
        },
        "nome": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "email": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "telefone": {
          "format": "int64",
          "maximum": 99999999999,
          "minimum": 11111111,
          "type": "integer"
        },
        "genero": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "nascimento": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        }
      }
    },
    "CreateCertificacaoCandidatoDTO": {
      "required": [
        "candidatoId",
        "certificacaoId"
      ],
      "type": "object",
      "properties": {
        "candidatoId": {
          "format": "int32",
          "type": "integer"
        },
        "certificacaoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateCertificacaoDTO": {
      "required": [
        "instituicao",
        "nomeCertificacao",
        "skillId"
      ],
      "type": "object",
      "properties": {
        "instituicao": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "nomeCertificacao": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "skillId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateSkillCandidatoDTO": {
      "required": [
        "candidatoId",
        "skillId"
      ],
      "type": "object",
      "properties": {
        "candidatoId": {
          "format": "int32",
          "type": "integer"
        },
        "skillId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CreateSkillDTO": {
      "required": [
        "nomeSkill"
      ],
      "type": "object",
      "properties": {
        "nomeSkill": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    }
  }
}